baccarat_bot/requirements.txt
# baccarat_bot/requirements.txt
# Dependencias del Bot de Baccarat

# Dependencias principales
python-telegram-bot==22.5
python-dotenv==1.2.1

# Dependencias para web scraping AGRESIVO y datos reales
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
playwright>=1.40.0
playwright-stealth>=0.1.5

# Dependencias para anti-bot evasion
undetected-chromedriver>=3.5.4
fake-useragent>=1.4.0

# Dependencias adicionales para manejo de datos y logging
requests>=2.28.0
psutil>=5.9.0

# Dependencias para el panel web y API
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.24.0

# Dependencias opcionales (para desarrollo)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=22.0.0
# flake8>=5.0.0
